Added Role checks for post editing/deleting on the Backend;

+++ server/src/controllers/post.controller.ts
@@ -25,25 +25,27 @@ export async function getPost(req: Request, res:Response) {
 }
 
 export async function deletePost(req: Request, res: Response) {
-  const post = await postService.getPostById(req.params.id);
+  if(!req.user) return res.status(401).json({message: "Unauthorized"})
 
+  const post = await postService.getPostById(req.params.id);
   if (!post) {
     return res.status(404).json({ message: "Post not found" });
   }
 
-  // if (
-  //   post.authorId !== req.user!.id &&
-  //   req.user!.role !== "ADMIN"
-  // ) {
-  //   return res.status(403).json({ message: "Forbidden" });
-  // }
+  const isOwner = post.authorId === req.user.id
+  const isAdmin = req.user.role === "ADMIN"
+  const isMod = req.user.role === "MOD"
+
+  if(!isOwner && !isAdmin && !isMod) {
+    return res.status(403).json({message: "Forbidden"})
+  }
 
   await postService.deletePost(post.id);
   res.status(204).send();
 }
 
 export async function updatePost(req: Request, res: Response) {
-  // if(!req.user) return res.status(401).json({message: "Unauthorized"})
+  if(!req.user) return res.status(401).json({message: "Unauthorized"})
 
   const {title, content} = req.body
   if(!title || !content) {
@@ -53,6 +55,13 @@ export async function updatePost(req: Request, res: Response) {
   const post = await postService.getPostById(req.params.id)
   if (!post) {
     return res.status(404).json({ message: "Post not found" });
+  }
+
+  const isOwner = post.authorId === req.user.id
+  const isAdmin = req.user.role === "ADMIN"
+
+  if(!isOwner && !isAdmin) {
+    return res.status(403).json({message: "Forbidden"})
   }
 
   const updatedPost = await postService.updatePost(post.id, title, content)
